Remove deprecated BookList component and use new dumb book list
