+++ assets/src/libraries/index.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import BookList from './BookList';
-import BookService from '../services/BookService';
+import Library from './Library';
 
 const container = document.getElementById('libraries');
 const library = container.getAttribute('data-library');
 
 ReactDOM.render((
-	<BookList service={new BookService()} librarySlug={library} />
+  <Library slug={library} />
 ), container);
